requirements.txt
streamlit
requests
python-dotenv
pandas
openpyxl
gtts
whisper
numpy
sounddevice
wavio
